improve readability of GameplayEngine.UpdateGameplay()
"Columns[i].UpdateHitObjects[k].Value" was being used over and over in this method and imo hurt readability. I assigned this to a variable, and referenced this variable within the method. From my testing, it doesn't hurt performance or cause allocations. In fact it may be more optimized than before, though my testing didn't conclude anything one way or the other.

There's probably other places where we could make changes similar to this. I just noticed this case last night, and wanted to see if there was a way to make it better.