Add a user-defined "Crosshair Offset", active with Hidden
Makes the crosshair appear at a different size than normal. The ZLocation and Diameter values of the crosshair do not change. Arcs are still "Perfect" at the same time and place because they see the map-intended size for the crosshair, not the size after offset.

Giving myself control of the crosshair during Hidden helped me read it better, so I think it's something to keep for now. There is other problems with Hidden to be addressed but this I think this is a temporary solution to address the fact that Hidden feels weird to play. Adjusting the crosshair helped me get dampen that feeling, let's see if it helps testers.

Negative values make the crosshair shrink, positive values make it grow. The number defined is how much the diameter (full width) of the crosshair will change in pixels on the base 1080p resolution. For example, if you give a value of -25, the crosshair's base diameter will become 275 instead of 300 pixels wide.