Add ACEEditor | add Editor singleton | misc reformattings
Due to the way I'm structuring this, an Editor singleton would be very useful for editors of multiple styles. Currently ACEEditor has to derive from the ACEngine, so I can't have an abstract parent that could cover all Editors. The singleton provides an answer for stuff that all Editors need, and so we don't need to copy/paste shit over and over for each new style.
It would also make swapping from one style to another much simpler as the new style can reference the singleton for the values that were being used in the previous style.