change starting delay for Gameplay
Used to force a 1 second wait on all maps, now it adds 100 ms if the first note is over 1 second away, and adds up to 1 second if the first note is within 1 second away.

Changed the delay to keep AudioRate in mind, I believe this fixed the weirdness that the start of maps showed during the delay, but idr if that was fixed before or not.